pyproject.toml
[tool.poetry]
name = "viz"
version = "0.1.4"
description = ""
authors = ["mmi <discoride.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "viz"}]

[tool.poetry.dependencies]
python = "3.11.2"
libvirt-python = "^10.7.0"
pika = "^1.3.2"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
devpi-client = "^7.1.0"
mypy = "^1.11.2"
watchdog = "^5.0.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
